Intermediate!
At the start of the game, Python’s random module is used to generate a random number within a specified range (e.g., 1 to 100). The player is then prompted to guess the number. If the player’s guess is too high or too low, the game provides a hint and allows the player to guess again. This process continues until the player guesses the correct number.